remove mongodb
