+++ src/contract/abi/SimpleAggregator.json
