+++ src/contract/util.ts
