+++ src/database/entity/BaseBlockEntity.ts
